backend/requirements.txt
alabaster==1.0.0
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.12.1
daphne==4.1.2
diff-match-patch==20241021
Django==5.1.3
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-countries==7.6.1
django-filter==24.3
django-graphql-jwt==0.4.0
django-import-export==4.3.3
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
docutils==0.21.2
easy-thumbnails==2.10
Faker==33.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.55.0
gevent==24.11.1
geventhttpclient==2.3.3
graphene==3.4.3
graphene-django==3.2.2
graphene-file-upload==1.2.0
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
locust==2.32.5
MarkupSafe==3.0.2
matplotlib==3.9.2
msgpack==1.1.0
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
pipdeptree==2.24.0
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyphen==0.17.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
# pypiwin32==223
pyzmq==26.2.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
#tablib==3.5.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validate_email==1.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
zope.event==5.0
zope.interface==7.2
zopfli==0.2.3.post1
